pyproject.toml
[tool.poetry]
name = "phillipa"
version = "0.1.0"
description = "Phillipa The Dolphin"
authors = ["Dan Trickey <dan@trickey.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = { git = "https://github.com/bryanforbes/discord.py.git", branch = "typings" }

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
flake8-commas = "^2.0"
flake8-comprehensions = "^3.2"
flake8-debugger = "^3.2"
flake8-docstrings = "^1.5"
flake8-isort = "^2.9"
flake8-mutable = "^1.2"
flake8-todo = "^0.7.0"
mypy = "^0.761"
rope = "^0.16.0"
isort = {version = "^4.3",extras = ["pyproject"]}


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
